Move shared testing code into datafusion_common
